fix docs generation
